+++ Model/Person.h
@@ -16,7 +16,7 @@ public:
     Person(std::string firstName, std::string lastName, double workHours);
     Person() = default;
     virtual ~Person() = default;
-    virtual long double calculateSalary() const = 0;   // pure virtual function
+    virtual  double calculateSalary() const = 0;   // pure virtual function
 };
